abstracts fetch request to reusable function #48
